requirements.txt
PyYAML
matplotlib
uvicorn
fastapi
pydantic
pandas
tqdm
numpy
scipy
TextGrid
pydantic